mcp_servers/elasticsearch_mcp/requirements.txt
elasticsearch>=7.0.0,<9.0.0
google-adk>=1.0.0
fastmcp>=0.1.0
httpx>=0.23.0
uvicorn>=0.18.0
python-dotenv>=1.0.0
aiohttp>=3.8.0